Add error handling to ns-build.sh
The bash script includes many parts that, if an error occurs, would cause the script to not fail correctly, or, even worse, leave the application in an unstable state.

This starts by introducing a cleanup function that runs on exit ('trap cleanup EXIT'), so that if the user interrupts the execution, the script correctly cleans up after itself.

Because the cleanup script has an expectation for the directory it runs in, it means the 'cd' commands could throw off its expected execution directory.  To this end, the 'cd' only runs in a sub-shell for the zip commands.

Next, most commands include an exit-on-failure ('|| exit 1').  This allows the script to better handle environment or build issues.

This also includes some extra cleanup.  The app-backup and nodejs and standalone directories are now put inside the new .ns-build directory.  This should help make .gitignore easier, and can help reduce root directory clutter and possible interference with the application structure.  Some bash-specific code changed to more sh-friendly syntax (such as '[[' to '['), and included extra argument quoting.

The one line that seems odd is the 'KEY="$(echo "$ARGUMENT" | cut -f1 -d=)" - the '-f1' would take all fields before the '-d=' mark, but I'm not sure if that's the intended behavior; should this instead be a '-f2-'?

With a bit of extra effort, the argument parsing loop could be turned into a more general SH script instead of bash-specific, but I'll leave to that to another day.