Return 404 for Non-Existent Prop
For the property construction, it's possible for either a bug or some other request to fetch a property file that does not exist.  The original code, if the page does not exist, would generate a 500 error and a stack trace.  This fix instead returns a 404 to indicate the requested resource does not exist.