Add unit tests to ns-build server
This allows constructing an environment to enable testing the server.ts file.  Because it requires a next.js built app to run, the unit tests require mocking up a built app.

The 'api-gateway-2.spec.ts' file shows how the tests can run.  The utility function creates the mock environment, loads the server file, and returns it for the tests to run.

While the code will work if the setup instead constructs the files under a temporary directory, it requires copying the server.[ts|js] files into the temporary directory.  This means the code now has added complexity for keeping track of all sources, and it means developers must debug with breakpoints on files that exist in temporary directories.